converts project to typescript

+++ src/services/authors/index.ts
@@ -1,15 +1,21 @@
-import express from 'express'
-import AuthorModel from './schema.js'
-import BlogPostsModel from '../blogPosts/schema.js'
+import express, { NextFunction, Request, Response } from 'express'
+import AuthorModel from './schema'
+import BlogPostsModel from '../blogPosts/schema'
 import createHttpError from 'http-errors'
-import { basicAuth } from '../../auth/basicAuth.js'
-import { adminAuth } from '../../auth/adminAuth.js'
+import { basicAuth } from '../../auth/basicAuth'
+import { adminAuth } from '../../auth/adminAuth'
+import { cloudinary, parser } from '../../utils/cloudinary'
 
 const authorsRouter = express.Router()
 
-authorsRouter.post('/', basicAuth, adminAuth, async (req, res, next) => {
+authorsRouter.post('/', parser.single('authorAvatar'), basicAuth, adminAuth, async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const newAuthor = new AuthorModel(req.body)
+        const { firstName, lastName } = req.body
+        const newAuthor = new AuthorModel({
+            ...req.body,
+            avatar: req.file?.path || `https://ui-avatars.com/api/?name=${firstName}+${lastName}`,
+            filename: req.file?.filename
+        })
         await newAuthor.save()
         res.status(201).send(newAuthor)
     } catch (error) {
@@ -34,10 +40,24 @@ authorsRouter.get('/me', basicAuth, async (req, res, next) => {
     }
 })
 
-authorsRouter.put('/me', basicAuth, async (req, res, next) => {
+authorsRouter.put('/me', parser.single('authorAvatar'), basicAuth, async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const editedAuthor = await AuthorModel.findByIdAndUpdate(req.author._id, req.body, { new: true })
-        editedAuthor ? res.send(editedAuthor) : next(createHttpError(404, `Author with id ${ req.author._id } not found.`))
+        if (req.author) {
+            const oldAuthor = await AuthorModel.findById(req.author._id)
+            if (oldAuthor) {
+                const body = { ...req.body, avatar: req.file?.path || oldAuthor.avatar, filename: req.file?.filename || oldAuthor.filename }
+                const editedAuthor = await AuthorModel.findByIdAndUpdate(req.author._id, body, { new: true })
+                if (!editedAuthor) return next(createHttpError(404, `Author with id ${req.author._id} not found.`))
+                if (oldAuthor && req.file) {
+                    await cloudinary.uploader.destroy(oldAuthor.filename)
+                }
+                res.send(editedAuthor)
+            } else {
+                next(createHttpError(404, `Author with id ${req.author._id} does not exist and cannot be edited.`))
+            }
+        } else {
+            next(createHttpError(400, 'Invalid request.'))
+        }
     } catch (error) {
         next(error)
     }
@@ -45,8 +65,16 @@ authorsRouter.put('/me', basicAuth, async (req, res, next) => {
 
 authorsRouter.delete('/me', basicAuth, async (req, res, next) => {
     try {
-        const deletedAuthor = await AuthorModel.findByIdAndDelete(req.author._id)
-        deletedAuthor ? res.status(204).send() : next(createHttpError(404, `Author with id ${ req.author._id } does not exist or had already been deleted.`))
+        if (req.author) {
+            const deletedAuthor = await AuthorModel.findByIdAndDelete(req.author._id)
+            if (!deletedAuthor) return next(createHttpError(404, `Author with id ${req.author._id} does not exist or had already been deleted.`))
+            if (deletedAuthor.filename) {
+                await cloudinary.uploader.destroy(deletedAuthor.filename)
+            }
+            res.status(204).send()
+        } else {
+            next(createHttpError(400, 'Invalid request.'))
+        }
     } catch (error) {
         next(error)
     }
@@ -54,8 +82,13 @@ authorsRouter.delete('/me', basicAuth, async (req, res, next) => {
 
 authorsRouter.get('/me/stories', basicAuth, async (req, res, next) => {
     try {
-        const authorPosts = await BlogPostsModel.find({ authors: req.author._id }).populate({ path: "authors", select: "firstName lastName" })
-        authorPosts ? res.send(authorPosts) : next(createHttpError(404, `Author with id ${ req.author._id } does not exist or had already been deleted.`))
+        if (req.author) {
+            const authorPosts = await BlogPostsModel.find({ authors: req.author._id }).populate({ path: "authors", select: "firstName lastName" })
+            authorPosts ? res.send(authorPosts) : next(createHttpError(404, `Author with id ${req.author._id} does not exist or had already been deleted.`))
+        }
+        else {
+            next(createHttpError(400, 'Invalid request.'))
+        }
     } catch (error) {
         next(error)
     }
@@ -67,20 +100,26 @@ authorsRouter.get('/:authorId', basicAuth, async (req, res, next) => {
         if (foundAuthor) {
             res.send(foundAuthor)
         } else {
-            next(createHttpError(404, `Author with id ${ req.params.authorId } does not exist or has been deleted.`))
+            next(createHttpError(404, `Author with id ${req.params.authorId} does not exist or has been deleted.`))
         }
     } catch (error) {
         next(error)
     }
 })
 
-authorsRouter.put('/:authorId', basicAuth, adminAuth, async (req, res, next) => {
+authorsRouter.put('/:authorId', parser.single('authorAvatar'), basicAuth, adminAuth, async (req, res, next) => {
     try {
-        const editedAuthor = await AuthorModel.findByIdAndUpdate(req.params.authorId, req.body, { new: true })
-        if (editedAuthor) {
+        const oldAuthor = await AuthorModel.findById(req.params.authorId)
+        if (oldAuthor) {
+            const body = { ...req.body, avatar: req.file?.path || oldAuthor.avatar, filename: req.file?.filename || oldAuthor.filename }
+            const editedAuthor = await AuthorModel.findByIdAndUpdate(req.params.authorId, body, { new: true })
+            if (!editedAuthor) return next(createHttpError(404, `Author with id ${req.params.authorId} not found.`))
+            if (oldAuthor && req.file) {
+                await cloudinary.uploader.destroy(oldAuthor.filename)
+            }
             res.send(editedAuthor)
         } else {
-            next(createHttpError(404, `Author with id ${ req.params.authorId } does not exist or has been deleted.`))
+            next(createHttpError(404, `Author with id ${req.params.authorId} does not exist and cannot be edited.`))
         }
     } catch (error) {
         next(error)
@@ -90,11 +129,11 @@ authorsRouter.put('/:authorId', basicAuth, adminAuth, async (req, res, next) => 
 authorsRouter.delete('/:authorId', basicAuth, adminAuth, async (req, res, next) => {
     try {
         const deletedAuthor = await AuthorModel.findByIdAndDelete(req.params.authorId)
-        if (deletedAuthor) {
-            res.status(204).send()
-        } else {
-            next(createHttpError(404, `Author with id ${ req.params.authorId } does not exist or had already been deleted.`))
+        if (!deletedAuthor) return next(createHttpError(404, `Author with id ${req.params.authorId} does not exist or had already been deleted.`))
+        if (deletedAuthor.filename) {
+            await cloudinary.uploader.destroy(deletedAuthor.filename)
         }
+        res.status(204).send()
     } catch (error) {
         next(error)
     }
